+++ jax/random.py
@@ -172,15 +172,12 @@ from jax._src.random import (
   randint as randint,
   random_gamma_p as random_gamma_p,
   rayleigh as rayleigh,
-  rbg_key as _deprecated_rbg_key,
   shuffle as _deprecated_shuffle,
   split as split,
   t as t,
-  threefry2x32_key as _deprecated_threefry2x32_key,
   triangular as triangular,
   truncated_normal as truncated_normal,
   uniform as uniform,
-  unsafe_rbg_key as _deprecated_unsafe_rbg_key,
   wald as wald,
   weibull_min as weibull_min,
   wrap_key_data as wrap_key_data,
@@ -191,36 +188,7 @@ from jax._src.prng import (
   threefry2x32_p as _deprecated_threefry2x32_p,
 )
 
-
-# Deprecations
-from jax._src.prng import PRNGKeyArray as _PRNGKeyArray
-
 _deprecations = {
-    # Added September 13, 2023:
-    "PRNGKeyArray": (
-        "jax.random.PRNGKeyArray is deprecated. Use jax.Array for annotations, and "
-        "jax.dtypes.issubdtype(arr.dtype, jax.dtypes.prng_key) for runtime detection of "
-        "typed prng keys (i.e. keys created with jax.random.key).\n"
-        "For more information, see https://jax.readthedocs.io/en/latest/jep/9263-typed-keys.html",
-        _PRNGKeyArray
-    ),
-    "KeyArray": (
-        "jax.random.KeyArray is deprecated. Use jax.Array for annotations, and "
-        "jax.dtypes.issubdtype(arr.dtype, jax.dtypes.prng_key) for runtime detection of "
-        "typed prng keys (i.e. keys created with jax.random.key).\n"
-        "For more information, see https://jax.readthedocs.io/en/latest/jep/9263-typed-keys.html",
-        _PRNGKeyArray
-    ),
-    # Added September 21, 2023
-    "threefry2x32_key": (
-        "jax.random.threefry2x32_key(seed) is deprecated. "
-        "Use jax.random.PRNGKey(seed, 'threefry2x32')", _deprecated_threefry2x32_key),
-    "rbg_key": (
-        "jax.random.rbg_key(seed) is deprecated. "
-        "Use jax.random.PRNGKey(seed, 'rbg')", _deprecated_rbg_key),
-    "unsafe_rbg_key": (
-        "jax.random.unsafe_rbg_key(seed) is deprecated. "
-        "Use jax.random.PRNGKey(seed, 'unsafe_rbg')", _deprecated_unsafe_rbg_key),
     # Added October 18, 2023
     "threefry_2x32": (  # Note: this has been raising a FutureWarning since 2021
         "jax.random.threefry_2x32 is deprecated. Use jax.extend.random.threefry_2x32.",
@@ -245,15 +213,10 @@ _deprecations = {
 
 import typing
 if typing.TYPE_CHECKING:
-  PRNGKeyArray = typing.Any
-  KeyArray = typing.Any
   default_prng_impl = _deprecated_default_prng_impl
   shuffle = _deprecated_shuffle
   threefry_2x32 = _deprecated_threefry_2x32
   threefry2x32_p = _deprecated_threefry2x32_p
-  threefry2x32_key = _deprecated_threefry2x32_key
-  rbg_key = _deprecated_rbg_key
-  unsafe_rbg_key = _deprecated_unsafe_rbg_key
 else:
   from jax._src.deprecations import deprecation_getattr as _deprecation_getattr
   __getattr__ = _deprecation_getattr(__name__, _deprecations)